Bugs Fixed
  1. Feature extractor weights never loaded - Added _load_connector_weights() method to load weights from models/LTX-2/connectors/diffusion_pytorch_model.safetensors
  2. Embeddings1DConnector missing - Created src/llm_dit/encoders/embeddings_connector.py with:
    - 2-layer bidirectional transformer with RoPE (split mode)
    - 128 learnable register tokens
    - Self-attention + feed-forward blocks
  3. RoPE batch size mismatch - Fixed _apply_split_rotary_emb() to use batch size from input_tensor instead of cos_freqs

+++ tests/e2e/test_pipeline_shapes.py
@@ -222,7 +222,7 @@ class TestPipelineShapesGPU:
         load_connector_weights(connector, connectors_path, prefix="video_connector.")
 
         additive_mask = (1.0 - attention_mask.float()) * -10000.0
-        additive_mask = additive_mask[:, None, None, :]
+        additive_mask = additive_mask[:, None, None, :].to(dtype)  # Match model dtype
 
         output, _ = connector(projected, additive_mask)
